+++ backend/jeecg-module-system/jeecg-system-biz/src/main/java/org/jeecg/modules/flowable/dto/FlowableProcessDefResp.java
@@ -21,4 +21,13 @@ public class FlowableProcessDefResp {
 
     @ApiModelProperty("is default")
     private Integer isDefault;
+
+    @ApiModelProperty("start permission code")
+    private String startPermCode;
+
+    @ApiModelProperty("can start")
+    private Boolean canStart;
+
+    @ApiModelProperty("missing permission code")
+    private String missingPerm;
 }