+++ backend/jeecg-module-system/jeecg-system-biz/src/main/java/org/jeecg/modules/flowable/service/IProcessRegistryService.java
@@ -17,4 +17,6 @@ public interface IProcessRegistryService {
     String resolveDefaultProcessKey(String formKey);
 
     FlowableProcFormBindResp getDefaultBindByProcessKey(String processKey);
+
+    String getStartPermCodeByProcessKey(String processKey);
 }